requirements.txt
# Function dependencies, for example:
# package>=version
sendgrid
google-cloud-datastore